Answer every domain operation with the handle it resolved
A caller could only name a domain by string, and the record walk answers with whatever carries that name now.
A slot that changed hands between a join and a leave therefore took the leave for the domain that replaced the one being left.
create, the name scan, join and leave now each answer or accept the handle that operation already resolved, and Peer refuses a slot whose incarnation moved on.
The paths that acted on a bare id are gone.

RequestAggPolicy::lock waits RequestWithdrawSettle after dropping demand, which RequestPolicy::lock already did.
Without the wait it answered NotArrived while the record still named the caller, and under agg that state does not clear on its own.
collectTransferableLocal runs on local belief the timed-out waiter never adopted, and transferHeldDomainsAgg grants only to a peer with demand raised.
Counting rounds where the record named the waiter after a timeout: request-agg was 11 of 679 and is 0 of 894 with the settle, at a cost of 10 us on the timeout path.
OrderPolicy fails the same property at a comparable rate and no settle can close it, since it grants unconditionally to the next active participant.

Three pieces moved into common/, since the library and the harness both read them.
The pause hint folds into the backoff, so a spin site states one policy instead of two.
The polling wait is one function rather than a loop written per caller.
A bitmap is named for its bit count alone, which is the one fact its owner and its reader share.

Widths follow whoever owns the value.
An integer takes the width the type that owns it decided, and a churn counter reaches printf through PRIu64 rather than a cast to %llu that states a width the code never chose.

Two cases now fail on what their names claim.
One asserted a constant, and the other only died at a boundary instead of holding it open.

+++ src/core/layout/geometry.hpp
@@ -10,7 +10,6 @@
 
 #pragma once
 
-#include <cstddef>
 #include <cstdint>
 #include <memory>
 #include <string_view>
@@ -123,7 +122,7 @@ public:
     struct DomainRecord_t
     {
         static constexpr std::uint32_t Magic = 0x434D4554u;  // "CMET"
-        static constexpr std::size_t MaxNameLen = MaxDomainNameLen;
+        static constexpr std::uint32_t MaxNameLen = MaxDomainNameLen;
 
         // Active = live domain; Free = reusable spare. createDomain fills Free then
         // flips Active as last write; a mid-write crash leaves it Free (safely reclaimable).
